// src/lib/common/types/ComponentTypes/GalleryType.ts
export interface GalleryPropsType extends GalleryStyledPropsType, GalleryItemWrapperStyledPropsType {
  children: React.ReactNode;
}

export interface GalleryStyledPropsType {
  id?: string;
  padding?: string;
}

export interface GalleryItemWrapperStyledPropsType {
  theme?: GalleryThemeType;
  gap?: GalleryGapType;
  column?: number;
}

type GalleryThemeType = 'mid-night' | 'blossom' | 'fruits' | 'bare-bare' | 'mint-chocolate';
type GalleryGapType = 'wider' | 'wide' | 'normal' | 'narrow' | 'narrower';